[-] fix skin changer [-] config sys reworked [-] fix compilation

+++ nullptr-recode/sdk/structures/c_econ_item_view.h
@@ -5,7 +5,7 @@ private:
 	using str_32 = char[32];
 public:
 	NETVAR(int32_t, initialized, "DT_BaseAttributableItem", "m_bInitialized");
-	NETVAR(int, item_definition_index, "DT_BaseAttributableItem", "m_iItemDefinitionIndex");
+	NETVAR(int16_t, item_definition_index, "DT_BaseAttributableItem", "m_iItemDefinitionIndex");
 	NETVAR(int32_t, entity_level, "DT_BaseAttributableItem", "m_iEntityLevel");
 	NETVAR(int32_t, account_id, "DT_BaseAttributableItem", "m_iAccountID");
 	NETVAR(int32_t, item_id_low, "DT_BaseAttributableItem", "m_iItemIDLow");
